PythonLearning/Roots/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
pydantic==1.10.15
httpx==0.27.0
python-dotenv==1.0.1
PyJWT==2.8.0


